get auth working so we can access a page
